Add section for jobs with specific build log msgs
There may be a specific build log message we are interested
in knowing all the jobs that contain it that doesn't
neccessarily show up in a junit test case. This commit
is adding the message 'Failed to pull image "busybox"' which
can help to know how prevalent the problem is where the
docker image pull rate limiting is causing image pull
failures. But, the logic could easily be refactored to
have a list of messages to search for and a simple loop
on those could produce however many sections of these
job groupings that would be useful.

Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>