Add image for build root and use latest for all images
In order to use ubi:8 as a build root, we need
git pre-installed per the CI documentation[1], as
well as any other deps needed to build the project.

There doesn't seem to be any ready-made image in
the cluster I can use for this, since we need
both npm and go.

This also updates the main Dockerfile to use
latest.

[1] https://docs.ci.openshift.org/docs/architecture/ci-operator/#build-root-image