Add react frontend for sippy
This adds a react frontend hosted at "/sippy-ng", with near exact
functionality as the current frontend, however not enirely complete.
The idea is the two can co-exist until it has full feature parity.

Details about the react app are in sippy-ng/README.md, including
about testing with Jest and Polly.