bugs: consider version when determining if bug is linked or not
The difference whether a bug is considered linked to a release, or only
associated is based on the target release. While a bug is untriaged,
target release is blank -- so bugs that are new but definitely linked to
the problem only show up as associated.

This considers the Version if target release is blank.