Fix sippy for an upstream testgrid change to Overall test name.
Late yesterday a prod push in testgrid rolled out a change which caused
the Overall test name to change to [jobName].Overall. As a result sippy
could no longer detect jobs completing and assumed everything was still
running.

This change adjusts our code to look for both formats when searching for
the Overall test, which is used to determine success/failure.