db sync: batch insertions and only fetch what is needed from bigquery
We're failing to insert job runs because the size of the insert is now
greater than 65,535 (max of int16). This changes the insert to go in
batches.

More importantly, we also only fetch what is exactly needed from
BigQuery.  If our DB is already populated, there's no need to pull
everything from BQ - this gets the highest ID from postgres, and fetches
only the newer records from BQ.

Note: BigQuery does not support autoincrementing primary keys, so we use
`ROW_NUMBER() OVER(...)` -- this produces stable ID's if we do not
remove older records. Currently, we use BQ as append-only and for the
foreseeable future will probably maintain release information
indefinitely (it's not a lot of data).