*: Add 4.11 to Sippy
Like 1b71c2d447 (Add 4.10 to Sippy, 2021-09-07, #283), but for 4.11.
In preparation for the pending branch cut in [1].

Diverging from 1b71c2d447, I've left
openshiftJobsNeverStableForVariants unpopulated for 4.11. Filling
those 4.11 entries in seems like a useful time for an audit to confirm
that the entries still have open bugs that weren't fixed in 4.10.

[1]: https://github.com/openshift/release/pull/25557