Stop stripping fixed list of test suites.
This list can't ever be kept up to date, and thus we end up with some
tests with a suite prefix and some without. Now that we're transitioning
to a db, this is not consistent enough for us to rely on.

We can't assume to split on "." because all kinds of tests are using
dots for a hierarchy, and renaming them would obsolete a bunch of data
we can't lose.

Include suite name in the test name.

In future if we move to be powered by junit files, or the bigquery
aggregation db, we can migrate this data.