Fix filtering for tests with trailing whitespace.
Some have snuck into origin, the whitespace is filtered out when we
build our filter string in the UI, resulting in no response when trying
to load the test details/analysis page in the UI.

Fix by trimming the test names in memory. Will go fix in master for
origin test names as well.