Add job runs endpoint using postgres DB
This adds a postgres-backed job runs API endpoint. Note that as part of
this change, there is an API change.  Job runs suffered from a mistake I
made early on, which resulted in mixed case within the API, and this
fixes it to be consistent (which allows wiring up the API endpoint with
SQL filtering).