Fix addition of synthetic tests for every job run.
Previously this was only added in some cases, openshift-tests should
work would be added only if we ran a test that looked like it came from
that suite. This code was new in the postgres work and made the
assumption that if we didn't mark the syntehtic test a fail, it must
have been a pass, when in fact it was intended to have a third didn't
run state.