Keep sippy operational when refreshing materialized views.
Our refreshes run for about 12 minutes total, some of them taking up to
5 minutes on a particular view. During this time reads were locked and
  we noticed pages stop responding in sippy.

Postgresql allows concurrent refreshes which do not lock the table under
some conditions. This PR adds the required unique indexes on each table,
this may actually speedup the app in general.

Concurrent refresh cannot be used on the first refresh, so we always
attempt it, if it fails we fall back to the regular refresh which will
be run on first startups for developers, or when we update the view
schema in a migration.

Added a test id column to one of the matviews that was using name so we
don't have to index on huge strings.

Updated to refresh metrics before and after matview refreshes as we
think this is causing gaps that triggers alerts to clear and resurface.