Remove in-memory sippy impl and the legacy API.
PostgreSQL sippy has been in the wild for some time now and running
well, we are ready to jettison all the old code.

Unfortunately required removing the server smoke test as we can no
longer run without a db, and we do not yet have a way to fire up a
temporary db for a server. This gap will need to be handled separately
but this did not catch much anyhow.

Includes removal of a duplicated testidentification directory that got
missed somehow.