Auto-filter test/job tables to those running at least once a day.
Currently this is only those with >= 1 current runs, resulting in the
first view you see being full of tests that ran only once and failed in
the last week. This is not really an area we focus on, moving up
shows us something much more high value initially, with the option to
adjust the filter lower if need be.