Guess which release to open in sidebar
Currently if you visit a 4.10 link, the 4.11 release is expanded in the
sidebar: https://sippy.dptools.openshift.org/sippy-ng/jobs/4.10

This tries to get the release from the URL and expand that one on first
loading of a link like the above. Using a React context would be better,
but unfortunately we're pretty deeply tied to passing release as a prop
(inaccessible to the sidebar).