Add download and copy names link
This adds two features:

1. Download

Enabled for both the jobs and tests page.

In the grid toolbar, a "Download" link is present that will give you the
raw JSON for the table rows.

2. Copy names

On the jobs table, add a "Copy names" button next to the analyzer button
to copy the selected job names (or all). This makes it easier to take a
filtered list of jobs and copy the names -- something I do manually
row-by-row by updating lists like never-stable.