Ensure we don't add duplicate PR's
The release controller has a bug when merging from upstream repos: it
assumes a mentioned PR belongs to the openshift fork, and makes a link
to it, despite the PR actually belonging to upstream.

If you look at
https://arm64.ocp.releases.ci.openshift.org/releasestream/4.12.0-0.nightly-arm64/release/4.12.0-0.nightly-arm64-2022-07-24-180825,
gcp-pd-csi-driver lists PR 1 multiple times, but only 1 refers to the
actual first PR to openshift, the other is from upstream.

Currently fetchdata is erroring out on ARM for this release payload,
this change prevents Sippy from trying to add the same PR twice. This
will just unbreak fetching data, but doesn't resolve the actual problem.

I filed a bug against the CRT team for the release-controller.

For Sippy, a proper fix is really moving away from the release pull
requests table and using the prow data -- we would then be able to
only adding PR to the release payload that we've seen in Prow.