Differentiate on test suite in NURPs test tables.
Often when hunting for tests that are degraded or worse than others, the
suite can make a dramatic difference in surfacing the problem, typically
during upgrade but not regular non-upgrade e2e.

Update the materialized view for test results to include suite name as a
characteristic.

Update the API to use this data when collapse=false, and ignore it when
collapse = true.

Similarly update the JS TestTable to display the column when
collapse=false.