Fix broken columns on release_job_runs.
This syntax resulted in a column named: prow_job_run_id,index:unique
We needed the semi-colon to tell gorm to separate the two, index and
unique are related thus the use of a comma, but if you're moving to a
different directive you  need to use semi-colon.

This error is completely silent because the old prow_job_run_id column
still exists, won't show up in prow or any dev setup that uses a prod
db. If you make a fresh db it will though.

We've got some bad data to cleanup since this change merged.