Show CI infra failures as overall status 'n'
Currently in prow-based Sippy, CI infrastructure failures are showing up
as a general non-e2e failure ('f'). This is because failed and errored
are considered the same. Instead, an errored prow job should be
considered a failure before setup (a.k.a. likely CI infra).

TestGrid sippy had no way to know this other than seeing a failed job
run without install results, which it called JobNoResults. This renames
it to more correctly be failure before setup (which is how the UI shows
it). Prow sippy can tell the difference.