Only update phase when change detected from release controller
If a release gets force accepted/rejected, the phase will change.
Previously we tried to refetch and regenerate the entire release tag
from the release controller when that happens, but the problem is Sippy
was re-creating all the job runs.  To avoid that, you would have to set
the ID on all the generated job runs to match what was in the DB, so
they would get updated and not created.

A simpler solution I believe is just to update the phase, as that's all
that's likely to change (and really all that we care about in TRT).