Very basic e2e tests for presubmit job.
Script that can be run by devs, and will be run by ci as a presubmit.
Script starts a docker container for postgres, loads the last 12 hours
of 4.7 data (takes about 5 mins for me), launches sippy API on a
non-standard port, and then runs a basic test that we can query job
data.

More tests will be added in future, but this alone will give us a big
jump in coverage for not breaking things.