Stop refreshing metrics early in startup.
We did this previously to prevent alert flapping but it did not work, we
now have a delay build into the promql queries in alerts.

Remove the initial refresh, and let the goroutine run RefreshData update
metrics after refreshing any matviews needing it. This doesn't happen
every time as we only refresh here if schema changed or the matview is
new.

Devs can start the server and the API is up quickly, even though we're
refreshing metrics and such in the background.