Remove build directory from the git repo
`sippy-ng/build` holds the artifacts from `make build`, so we can
produce an all-in-one binary used in production. `go:embed` embeds the
build dir in the sippy binary but freaks out if the directory is empty.
So we have to have something.

I originally put an index.html so you get something helpful to
understand why the all-in-one binary wasn't working, but it causes tons
of problems since it's the same file that `make build` writes to and
people keep overwriting it.

You can't ignore changes with `.gitignore` to a file that's already in
the repo.

Therefore, this deletes the build directory and creates it with a
Makefile.  A user trying to do `go build .` directly will need to make
the directory themselves.