Only call GitHub once per PR
We track PR's based off the PR and the HEAD SHA.  There's a bug where we
query GitHub for each sha, instead of just once per PR. This adjusts
the caching and syncing logic to ensure we only git the API once for a
PR.

GitHub limits us to 5,000 requests per hour. We know more than that now,
so we're hitting the limits.