Cascade delete for all relevant DB models
When deleting a DB row that has relations, these are not automatically
deleted and the foreign key constraints prevent you from actually
deleting the row. We need to cascade the deletes, however gorm does not
support patching constraints[1], therefore we have to delete them.

There's a one-time SQL script in scripts/ that we can run to clear the
constraints, which gorm wil re-add back. I've tested this on a
production database restore and it seems to do what I expect. We could
write our own migrator, but it looks somewhat challenging, and this is
probably a one-ish time thing.

[1] https://github.com/go-gorm/gorm/issues/5559#issuecomment-1200718658