Use consistent cache key for tests that includes suite name
TRT-593

We're sometimes marking a test incorrectly as a flake. This is because
of the suite handling in Sippy.  We got 1 test pass in one suite, and a
failure in separate suite, and sippy was combining them. This is a
regression, we previously fixed it but it got reintroduced in juggling
suite names.

Now that we always know the suite since we have access to the raw junit,
maybe we stop with only sometimes prepending it? But for now this fixes
the cache to always include the suite so we don't accidentally combine
results from different suites..