Estimate changepoint for job analysis
Changepoint analysis allows us to detect roughly when a change occured,
which is not always simple to spot for an unreliable test.

This uses a go package to do changepoint analysis on our job pass rates,
plotting a dashed white line on the job analysis on our best guesses
when the job changed state.