Fix or filtering
Wasn't working as we inject an AND for the release first, then we try to
add ORs but of course the AND already picks up all tests, so ORs had no
effect.

This makes it possible to OR on test names. However it still doesn't
work if you use any filtering on fields beyond the rawFilter, as we
split into raw and post filtering which comes after.