Init database schema before each deployment
Pods no longer try to update schema on startup.
Fetchdata WILL continue to update schema before it does it's work. This
should be safer as fetchdata runs on a schedule that is disjoint from
the pods rolling out new revisions. It might occasionally go wrong but
we will then retry.