Refactor db load error handling to tolerate problems and keep going
This is to hopefully allow more resilient syncing without abandoning
large portions because of a single area in any specific section of the
code.

All errors are re-logged at the end of the process, and if any were
present an overall error is returned causing the job to fail. This will
hopefully fire alerts, and the logs at the end of the pod should
summarize what went wrong.