Only render testgrid link when present
Currently we always render the TestGrid button regardless of whether
Sippy knows the testgrid link for that job.  This change makes it only
render when we know the URL. This will make it easier to identify jobs
where we don't know the TG dashboard, as well as not showing on jobs
that will never have one (e.g. presubmits).