Add infrastructure success metrics
This adds new metrics for per-platform infrastructure success rates.
This is a global value not constrained by releases, but we could scope
it by release if we thought it was useful.

This metric is the true success rate, so alerts based on this value will
be taking into account both failures and flakes (i.e. retrying infra
worked). This will let us make sure that we don't excessively start
flaking on infra and not detect the problem.