Change how aggregated variants are filtered out
This moves the filtering out of aggregation from test results from being
hardcoded to being done in the UI.  By moving it to the UI, the filter
can be removed and you can view tests for only aggregation. We need to
do this because there are tests that *only* appear in aggregation (e.g.
the new disruption tests).

The reason for needing a default exclusion is because the jobs that make
up an aggregated run are already tracked in Sippy. We are injecting
unwanted hysteresis that will cause changes to pass rates to lag behind.
The effect is probably small. Here's example example:

The follow test runs 10x and passes 7.

```
[sig-pizza] Pizza should be delivered within 30 minutes
```

It has a pass rate of 70%, in line with historical results.

If it is aggregated, we have an 11th entry in Sippy of a pass. Sippy
actually thinks the test passed 8 times out of 11 -- a pass rate of
72.7%.