Add an API for fetching job run intervals
This will hopefully soon be used by a sippy UI similar to the current
spyglass charts.

For now we fetch from GCS. In future, we may be pulling from loki or
opensearch.

To work around the fact there are often two intervals files (upgrade and
conformance phases), for now we return both added together with a
filename attached to each.