Fix installation logic determination
We were not correctly excluding upgrade-minor and aggregated from
install statistics everywhere.

- upgrade-minor installs the previous release and should not be counted.

- The reason for excluding aggregated is a little more subtle, but it
injects unwanted hysteresis that will cause changes to pass rates to
lag behind if we do not filter it out.

Here's an example:

The follow test runs 10x and passes 7.

    [sig-pizza] Pizza should be delivered within 30 minutes

It has a pass rate of 70%, in line with historical results.

If it is aggregated, we have an 11th entry in Sippy of a pass. Sippy actually thinks the test passed 8 times out of 11 -- a pass rate of 72.7%.