Enable batching for query params (back button)
Not a complete fix for the back button, but partially fixes it. One
issue is successive calls to set query params each create a history
entry. This turns on batching which makes consecutive calls only affect
the URL once.  Not all our calls are consecutive so we'd have to fix
that to get the back button working completely I think.