Skip bug lookup in e2e
Bug lookup hits search.ci, and it's quite an intensive operation. With
the quantity of PR's we landed today, I think we're hammering it a bit
too hard and it's affecting production Sippy's bug syncing.

We can re-instate this once we figure out how to reduce our load on
search.ci, there's an existing jira card for that.