Count test failures only once per job
Some test suites in OpenShift retry, resulting in potentially mulitple
failures for the same test in a job.  Component Readiness is currently
counting these as separate failures, resulting in an outsized impact on
our statistical analysis.

Counting them only once with the schema we have is challenging, this
creates a couple of CTE to split the table into failed test results,
and non-failed (success and flake).  We partition the failed test
results by name and prow job ID, and only select the first row from
each partition and then union it all with the non-failed table.