Show oldest date for resolution in jira incident calendar
OCPBUGS is a little weird, we don't get a resolution date until GA when
the bug is formally closed (or it's otherwise closed as a dupe, etc). We
need to find the changelog history of the first terminal state change,
i.e. the first date of state change to MODIFIED, ON_QA, Verified, or
Closed (in that order).