Ensure stale issues get updated for incidents
If we removed the incident label from a jira, we want to make sure we
update the database row still. So, we keep track of jira incidents we
didn't see yet and then query jira to figure out what happened. If the
jira got the trt-incident label removed, we found out when that happened
and set it as the incident resolution date.