Avoid gorm bug with smallint postgres column
Gorm appears to have a bug with detecting smallint/tinyint columns, it
thinks they always need their schema updated, whereas using an
int/bigint does not hit this.

Combined with a rollout of the deployment during our long running backup
job, this deadlocks the db somehow, but we need to remember this could
happen on any schema change if it ran into our backup in the middle of
the night.