Publish disruption delta metrics
Query new views in BigQuery which contain the complex logic, whatever is
in those views, we publish a metric for. One of them will need to be
updated as part of the GA date tasks.

Metrics published for the P50/75/95 delta now compared to both 30 days
prior to previous GA, and two weeks prior to two weeks ago.

Also includes the delta in PercentageAboveZero, which looks to volatile
to alert on now but we'll see if we can get a handle on it.

We'll alert on these to notify ourselves of regressions in either.