Publish disruption relevance metrics
This will allow us to join on the disruption metric and the relevance
metric to only alert on nurps we feel are relevant, as defined by the
view we're querying.

promql I tested with this looks like:

sippy_disruption_vs_prev_ga{delta="P50",release="4.15"} >= 1 and ignoring(delta) sippy_disruption_vs_prev_ga_relevance >= 2

It is returning the same as my sql queries so I believe the join is
working properly.