Track regression opened/closed dates in the metrics loop
When sippy issues the default component readiness query for the purpose
of populating component readiness metrics, use the report we just
generated to maintain a new test_regressions table. First time we see a
regressed test on a particular set of variants, we insert into the
table. As soon as it disappears from the report, we mark it closed.