Ignore disruption intervals without the display flag.
Small hack to work around problems moving towards a UI that only shows
intervals if the code generating them said they were meant for user
display, which is not presently in all releases we're using. As such,
overriding the display flag right now is very useful for 4.16, but this
includes some disruption intervals we should never be showing. I should
have separated them on Source, but again, hard to get that to all active
releases.

For now, have the UI never display those so the "override display flag"
works well on 4.16.