Risk Analysis: ensure Test ID is populated
At the time that openshift-e2e-tests requests RA, test results haven't
yet been associated with test IDs (we just have junits). So it is
posting a `ProwJobRun` with partial `Test` records to the API, sans IDs.
This PR changes the past-results queries by test name to also return the
test ID, and uses that to populate RA results.