Introduce lag for a regression to clear
This is an attempt to solve the problem with regressions flapping, if we
require 95% confidence in a regression before we mark you red, we do not
want components then immediately going green at 94.5% certainty because
job runs changed slightly.

Use the test_regressions tracking table (assume for default view now,
soon to be per-view as desired once server side views land). If we see
an open regression for a test, the required confidence for red becomes
confidence - 5 (as implemented). i.e. with the default 95% confidence,
that drops to 90% confidence of a regression if the test has an open
regression.

All of this takes effect in the code tracking open regressions as well,
so once the test is below 90% it will then get closed in the tracking
table, and now the component is green again.