cross-variant comparison: UI inputs
Enable the UI for specifying variants for cross-comparison, both in the
UI parameters and API parameters. This introduces the "Compare" button
in the variant selection sidebar, and plumbs the results through. This
adds parameters for "variantCrossCompare" and "compareVariant" which
specify (respectively) the variant groups that should query for
different values in sample vs basis, and what those actual values should
be.

There was also a fair amount of refactoring and renaming along the way.
I tried to distinguish variantGroup ("Architecture") from individual
variants ("Architecture:amd64"). Hopefully the result is clearer.