TRT-1813: fix bug omitting some variants from test_details filter
This was introduced with the variant cross-compare; originally the
CompareVariants would only include variants under cross-compare, so I
used that to determine whether this bit of code needed to add the
variant filter or not. Later I updated the code to fill out
CompareVariants with all the variants, which caused this bit to skip
adding basically everything. Thus Owner, which is not in the
RequestedVariants, if not in the cross-comparison would never be
filtered. By scoping the exclusion to the variant groups under
cross-comparison, it returns to its original intent.