Move regression tracking to fetchdata
This was done in the sippy pods themselves as they already query the
component report, so it seemed most efficient.

However we use caching, and with this method multiple pod replicas were
trying to do the same thing and it appears they sometimes conflict.

Simplify by doing this in a single thread when we run fetchdata every
few hours.