Hacky solution to test rarely run jobs results
Getting this to work fully in some kind of generic fashion across the UI
and query params will be extremely difficult. We're not even sure the
idea will work, so for now this PR adds a hacky temporary solution.

A scheduled query is doing a full table scan every day and copying
rarely run junit results to a much smaller table in bigquery,
junit_rarely_run_jobs. This allows us to scan it without the immense
cost.

For this temporary view, I added a hack where if we see you're querying
for JobTier rare, we will flip to using this other table.

Side effect, if someone uses the UI and adds in this JobTier, suddenly
their results will be fully busted.