Allow page loads when test analysis is empty
This is working around an edge case with our new method of daily
reporting on test analysis via bigquery, rather than matview refreshes.
A few tests exist in sippy but are not imported into bigquery, and as
such no longer have any test analysis available. This was causing the
page to crash.

Show a clear indicator that no analysis is available and keep the page
loading properly for the rest of it.