Log fatal and exit on redis outage
We believe specifying a redis url when no redis is listening is linked
to cost issues for developers using the views file, which is very easy
to do. Redis outages would be rare, so we're now going to exit if we
detect a redis url specified, but get a connection refused.