Require special approval for intentional regs in consecutive releases.
Surface the problem via a unit test which will fail PRs adding an
intentional regression for two releases in a row, for the same test.

To work past this, a new overrides.json file is added in a new directory
with very limited approvers. Test output directs to a README in this
directory signalling how to add to the overrides.json, and reach out to
get approval.