Negative binomial PR analysis
The R version came first and essentially uses the dnbinom function.  I
had AI translate it into Go and it had to do some of the math the hard
way.  The Go output is nearly identical, just lacking some precision.

An idea would be to run trials with various distributions to see which
turned out to be the most accurate.

What matters in this code is the "Posterior Predictive Probability"
output.  A low probabily means the prior data doesn't fit the new data
well so it might be worth investigating.  You can see that setting a
threshold of "Flag the test if we're 95% certain something weird is
going on" would work in all these examples.  In practice, I think we
could set it to 99% (Meaning, the Posterior Predictive Probability would
be <= 1%).

Right now this doesn't feel that different than p-values and Fisher's
Exact Number, but I think the nice part is that the math can work with
any sample size at all---even a sample of 1.  Obviously, the more data
that comes in the more confidence we'd have.