Refactor variant registry code for job tiers
This is a large refactor of the variant registry code -- no more
regexes, and it should be more of a "one stop shop" to change something,
since all the logic is held in each setter.

- Introduces an "excluded" job tier, used for anything not in Sippy's
  config file
- Also marks non-automated-release QE jobs as excluded

This can eventually be used to match Sippy core and CR jobs, but I want
the variants applied before seeing how component readiness starts
looking.