Attempt a fix for postgres CPU
In working on reducing matview CPU, I recently moved us to a partitioned
table test_analysis_by_job_by_dates, updated once a day, and we do not
prune old dates as we intended to allow configurable ranges to generate
the test pass rate graphs.

I analyzed running queries on postgres and found many risk analysis
queries, tracing back to this use of the new table without a filter on
dates. This means jobs are constantly running and requesting risk
analysis, and scanning this table that's growing larger and larger
without a date filter. Fixing this may well fix our problem with the
matviews as this could be what's taking all the cpu.