Show feature gates with zero tests
This adds a feature that syncs the list of feature gates from the
openshift/api repo, and tracks where they are enabled. This also allows
us to report which feature gates have 0 tests.

The mechanism use is to clone the openshift/api repo, and examine the
YAML manifest for each of the releases we're interested in (everything
4.15+ has the manifests).