Track the max number of regression failures
For the duration a regression is open, track the max number of failures we saw in the sample window. This will be helpful for determining what minFail we should use, and provide insights related to regressions that go away on their own.

Doesn't work great with triage right now for the same reason we see 100% passing on a triaged test details page, we decrement traiged runs today and don't return any info on them. This will be better in the new traige where job runs only get decremented when the regression is considered closed.

Also removed temp bigquery migration code for the postgres rollout that
I forgot to come back for.