jobartifacts: cache jobrun results
Job run results are now cached in the Redis cache for 4 hours.
Each query is cached by individual run.
If a result included any timed-out artifacts, when the cached entry is
returned, it will attempt to query the missing artifacts again so that
the result improves as it is retried.

For easier deserialization, MatchedContent became an actual type instead
of interface{} - should be clearer anyway.