Fix last failure time picking up flakes
This looks to have come in with the non-enabled treat flakes as failures
feature, but we mistakenly assumed to do that all the time in the query
for calculating last failure time. As a result, we show tests that have
flaked since their triage resolution as pants on fire when we didn't
actually count a test failure.

Last failure will now be correct for the main use case, but not for
flakes as failures, we'll need to revisit if that feature ever takes off
in a default view.

I've also switched to using prowjob start time for the last failure, not
modified time which I believe is ingestion time. This was confusing as
the test details page shows job runs and their start times.