[TRT-2044] bugloader: associate mapped tests with bugs
On test_details pages we look up bugs that have been linked according to
the test name. However sometimes test names change and are mapped to a
shared "unique" id. In this case, we would like the bug to be linked not
only from the test_details for the test name it mentions, but also from
the test_details of any test mapped with that.

This update enhances the bugloader to look up mapped tests as bugs are
scanned and link those in bug_tests as well as the mentioned test.