Move regression/triage data outside caching
This prevents a problem where a user updating triage data invalidates
the entire cache. We were using a ReportModified timestamp that would be
updated every time we changed triage data. This would cause a cache miss
on the report, but also the expensive underlying bigquery queries which
was done in one shot for all regressed tests, which we cannot recreate
here.

Injection of regression/triage data is fast and from local postgres, so
this is now moved outside the main report caching and PostAnalysis is
run on the report just before we return it. This works fine for current
use cases. If this does get too slow in the future, we can also cache it
here as well, but separately so we do not have to expire our expensive
keys.