Hide triaged resolved ci-infra regressions
They will reappear if failures are observed after the resolution date.

Should be visible in the new All Triages page, though not with
regressions attached as those would be gone.

This also means the regression tracker will close the regression at that
time, but it will also reuse the old regression record if it reopens.