backend: refactor PreviousRelease
There are a few places that actually need to know what release precedes
a given release, and for that, instead of parsing out a version and
decrementing, just plumb the release configs to it and follow the
PreviousRelease setting. This should work for OKD or whatever comes
along that has a concept of release progression no matter what the names
look like.