CR UI: set default sample release correctly
At init, there is code to determine from the list of releases what
should be the default query. We never actually use this anymore, but it
would be used if we did not have any views defined.

However it showed up in a bug where, at load, fetchData ran before
initializing the view, and default sample was set to an OKD release
interleaved in the list, resulting in a failed request that caused the
page to explode (fortunate, really, since otherwise it would make an
expensive query before throwing away the result and making the correct
query). It seemed wise to make this correct (even though the bug is
fixed and this won't get used).